03. 學生成績預測/requirements.txt
plotly~=4.14.1
pandas==1.1.2
streamlit==0.73.1
numpy==1.19.3
matplotlib~=3.3.3

scikit-learn~=0.24.0
tensorflow-cpu~=2.3.0
python-docx~=0.8.10
base58
xlsxwriter~=1.3.7
git+https://github.com/tensorflow/docs